Refactored how alerts are loaded.
Previously, alerts were loaded such that some data was ignored and some alerts were stored multiple times.  When an alert is for multiple geographic places, it occurs multiple times in HealthMap.  We stored the alert with an internal name of (HealthMap Id)(4 digit number) where the 4 digit number was based on how it was grouped in the response from HealthMap.  If this internal name already existed, the later alert would replace an earlier one.

I've changed this so that the alert is always stored with the internal name of (HealthMap Id)0000.  Within this record, all different place data is stored in the metadata under a 'places' list.  Arbitrarily, whichever place has the lowest place_id is stored in the main metadata.  The means that existing interfaces will continue to work with this data, even if they are not aware of the places list.

When updating a data set that was generated in the old method, existing items are renamed or consolidated only if they are processed within the new feed.  If there was no item with an internal name of (HealthMap Id)0000 and there is an existing entry for the item, one of the existing entries is renamed to (HealthMap Id)0000.  All entries with the same HealthMap Id and a non-zero 4 digit suffix are removed.

This requires a new version of the gritsSearch plugin to take advantage of the new data layout.

Additionally, I added a timezone library function to ensure that the date on the alert was properly being stored or compared with UTC dates and times.