Many changes to loading alerts; refactoring.
I broke apart the main function to several subfunctions to improve code readabily and make it easier to maintain.

The 'places' list has been renamed 'events'.  This is defined in a module constant in case we change our minds.

If an alert ever occurs without a place id, a warning will be issued.

If a place entry for an alert goes away (probably because the place was changed), and we have enough information to be certain that this change is certain, delete the place that is no longer available from the alert.

I added a fair amount of documentation.